requirements.txt
Django==6.0.3
djangorestframework==3.17.1
psycopg2-binary==2.9.11
django-allauth==65.15.1
reportlab==4.4.10
Pillow==12.2.0
django-widget-tweaks==1.5.1
django-redis==5.4.0
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0